Addons validation barely passed (false positive on front controller, regex is too long) Added information for shortcode use on configuration file Still fr translations

+++ classes/controller/FrontController.php
@@ -190,7 +190,11 @@ class EverPsBlogModuleFrontController extends ModuleFrontController
         $totalItems = (int)$total;
         $page = (int)Tools::getValue('page');
         $page = (int)Tools::getValue('page') ? (int)Tools::getValue('page') : 1;
-        $totalPerPage = (int)Configuration::get('EVERPSBLOG_PAGINATION') ? (int)(int)Configuration::get('EVERPSBLOG_PAGINATION') : 10;
+        $totalPerPage = (int)Configuration::get(
+            'EVERPSBLOG_PAGINATION'
+        ) ? (int)Configuration::get(
+            'EVERPSBLOG_PAGINATION'
+        ) : 10;
         $pagination = new Pagination();
         $pagination
             ->setPage($page)
